feat: add integration test
A bare smoke test, trying some simple variants on how a struct could be
used. Uses clang to compile the generated C code, and checks if clang
has any diagnostics about it.

Actual functionality of the getter/setter is not tested, but at least we
know the generated code compiles.

Also add a logic to pick up environment variable based clang args in the
actual generator.

Signed-off-by: wucke13 <wucke13+github@gmail.com>