fix: avoid UB due to creation of unaligned pointer
The previous code would happily cast `void*` with arbitrary alignment
to e.g. `uint64_t`. That is UB.

To fix this, instead `memcpy` to a temporary, but aligned instance of
the target type in question (e.g. `uint64_t`).

Reworks the array logic as well to cater to arrays of multibyte sized
elements.

Also alter the smoke_test to include both an array of 1 byte sized type,
but crucially also an array of multibyte sized elements.

Signed-off-by: wucke13 <wucke13+github@gmail.com>