feat: add secondary stack & TLS alloc functions
Adds some C code that should approximate the needed steps to set up a
linear stack and TLS for a new thread.

For now, it is appended as suffix to the generated `ARINC653-wasm.h`.
As the code needs to know `SYSTEM_LIMIT_NUMBER_OF_PROCESSES` which gets
its default value in `ARINC653-wasm.h`, we can only emit this code
after that header file was processed (so that the `SYSTEM_LIMIT_*`
defines are already known).

Signed-off-by: wucke13 <wucke13+github@gmail.com>