refactor: rewrite apex_wasm_proc_alloc in asm 2/2
Rewriting the function in assembly allows us to tightly control whether
the secondary stack is used. No local variables must be declared in the
function, as soon as they exist, at least in debug build code accessing
the linear stack is emitted.

Expressing the atomic synchronisation purely in inline assembly
alleviates the need for the inclusion of `stdatomic.h`. However, for the
`offsetof` macro, `stddef.h` is still needed.

Signed-off-by: Wanja Zaeske <wanja.zaeske@dlr.de>